moved FlightInfoFrame class and BookingFrame class into its own java file
